V2 of the Piezo Buzzer
Piezo buzzer is now utilizing the PWM for output compare to drive the Piezo buzzer. We are no longer trying to manually toggle using the Timer 2 interrupt. We also moved the pin of the Piezo buzzer RP11 (physical pin 22) on the PIC24, as there were previous pin conflicts with the stepper motor library. In the wiring of the standalone Piezo Buzzer unit a status LED was added to ensure that the PIC 24 is receiving power. Also updated the main testing file (dehm0036_piezoMain.c) to include a testing of the tones, and varying the frequency of the tone, and including a simple scale.