First Hot Potato Schematic Upload
added all components for the Hot Potato game to the schematic. I was unable to find the Neo-Pixel part for the software being used (Fritzing) So I created it from scratch as seen on the bottom right of the schematic. The other parts of the schematic include the stepper motor as well as it's driver board. There is also a start button that will run the main game function connected from RP8 to GND. The buzzer was moved to RP11 as there was a previous pin conflict.

The power module was also added that includes 4 AA batteries wired in series that should produce around 6V. 6V will be used to power the stepper motor, while the remaining circuit and components are powered off of around 3.15V after the 6V power is fed through a circuit divider.