Uploading NEOPIXEL Code
I included the header and source file, as well as the main file for testing. 
The header file contains declarations for functions write_0, write_1, delay_100micro, delay_1milli, and delay, as well as code for writeColor.
The assembly file contains the assembly code implementation for write_0, write_1, delay_100micro, delay_1milli. 
The main C file configures the PIC24 microcontroller's settings, initializes Timer3, and defines the main() function. The main() function repeatedly cycles through a loop, changing the color of the LEDs, as well as the frequency of the flashing, based on how close the game is getting to the end.